- pause sreen is full screen - can change keybind on pause screen - use PauseControl.cs to pause instead of PopOptionMane.cs - updated SpawnByShortcut.cs so that user cannot spawn when game is paused
