+++ Assets/Game/Scripts/PauseControl.cs
@@ -1,0 +1,35 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PauseControl : MonoBehaviour
+{
+    public static bool gameIsPaused; 
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.Escape))
+        {
+            gameIsPaused = !gameIsPaused;
+            PauseGame(); 
+        }
+    }
+    void PauseGame()
+    {
+        if (gameIsPaused)
+        {
+            Time.timeScale = 0f; 
+        }
+        else
+        {
+            Time.timeScale = 1; 
+        }
+    }
+
+    public void resumeGame()
+    {
+        gameIsPaused = false;
+        PauseGame(); 
+    }
+}
