+++ Assets/Game/Scripts/SpawnByShortcut.cs
@@ -14,29 +14,31 @@ public class SpawnByShortcut : MonoBehaviour
 
     void Update()
     {
-
-        if (contentToSpawn != null && !Input.GetKeyDown(KeyStrokeUtil.GetKeyStroke(contentToSpawn.PlayerPrefKey, contentToSpawn.DefaultKeyStroke)))
+        if (!PauseControl.gameIsPaused)
         {
-            contentToSpawn = null;
-        }
-
-        if (contentToSpawn == null)
-        {
-            foreach (SpawnContentByShortcut spawnContent in spawnContents)
+            if (contentToSpawn != null && !Input.GetKeyDown(KeyStrokeUtil.GetKeyStroke(contentToSpawn.PlayerPrefKey, contentToSpawn.DefaultKeyStroke)))
             {
-                spawnContent.timePassedDoNotTouch -= Time.deltaTime;
-                if (Input.GetKeyDown(KeyStrokeUtil.GetKeyStroke(spawnContent.PlayerPrefKey, spawnContent.DefaultKeyStroke)) && player.GetRewardAmountCollected() >= spawnContent.RewardCost && spawnContent.timePassedDoNotTouch <= 0f)
+                contentToSpawn = null;
+            }
+
+            if (contentToSpawn == null)
+            {
+                foreach (SpawnContentByShortcut spawnContent in spawnContents)
                 {
-                    contentToSpawn = spawnContent;
-                    spawnContent.timePassedDoNotTouch = spawnContent.CoolDown;
-                    GameObject newInstance = PoolManager.Spawn(contentToSpawn.Prefab, player.GetMover().GetNavMeshAgent().destination);
-                    if (!String.IsNullOrEmpty(contentToSpawn.SpawnSound))
+                    spawnContent.timePassedDoNotTouch -= Time.deltaTime;
+                    if (Input.GetKeyDown(KeyStrokeUtil.GetKeyStroke(spawnContent.PlayerPrefKey, spawnContent.DefaultKeyStroke)) && player.GetRewardAmountCollected() >= spawnContent.RewardCost && spawnContent.timePassedDoNotTouch <= 0f)
                     {
-                        FMODUnity.RuntimeManager.PlayOneShotAttached(contentToSpawn.SpawnSound, newInstance);
+                        contentToSpawn = spawnContent;
+                        spawnContent.timePassedDoNotTouch = spawnContent.CoolDown;
+                        GameObject newInstance = PoolManager.Spawn(contentToSpawn.Prefab, player.GetMover().GetNavMeshAgent().destination);
+                        if (!String.IsNullOrEmpty(contentToSpawn.SpawnSound))
+                        {
+                            FMODUnity.RuntimeManager.PlayOneShotAttached(contentToSpawn.SpawnSound, newInstance);
+                        }
+                        player.DecrementRewardCollected(contentToSpawn.RewardCost);
                     }
-                    player.DecrementRewardCollected(contentToSpawn.RewardCost);
                 }
             }
-        }
+        } 
     }
 }
